Factor out crossinterp_data_lookup.h.
